CVSL-3056 Moved the HasElectronicMonitoringResponseProvider to the correct package
